[WIP] Initial refactoring.
